+++ src/Paystack.php
@@ -100,41 +100,54 @@ class Paystack
         );
     }
 
-    /**
+
+     /**
+
      * Initiate a payment request to Paystack
-     * Included the option to pass the payload to this method for situations 
+     * Included the option to pass the payload to this method for situations
      * when the payload is built on the fly (not passed to the controller from a view)
      * @return Paystack
      */
+
     public function makePaymentRequest( $data = null)
     {
         if ( $data == null ) {
-            $data = [
-                "amount" => intval(request()->amount),
+
+            $quantity = intval(request()->quantity ?? 1);
+
+            $data = array_filter([
+                "amount" => intval(request()->amount) * $quantity,
                 "reference" => request()->reference,
                 "email" => request()->email,
                 "plan" => request()->plan,
                 "first_name" => request()->first_name,
                 "last_name" => request()->last_name,
                 "callback_url" => request()->callback_url,
+                "currency" => (request()->currency != ""  ? request()->currency : "NGN"),
+                
+                /*
+                    Paystack allows for transactions to be split into a subaccount -
+                    The following lines trap the subaccount ID - as well as the ammount to charge the subaccount (if overriden in the form)
+                    both values need to be entered within hidden input fields
+                */
+                "subaccount" => request()->subaccount,
+                "transaction_charge" => request()->transaction_charge,
+                
                 /*
                 * to allow use of metadata on Paystack dashboard and a means to return additional data back to redirect url
                 * form need an input field: <input type="hidden" name="metadata" value="{{ json_encode($array) }}" >
-                *array must be set up as: $array = [ 'custom_fields' => [
-                *                                                            ['display_name' => "Cart Id", "variable_name" => "cart_id", "value" => "2"],
-                *                                                            ['display_name' => "Sex", "variable_name" => "sex", "value" => "female"],
-                *                                                            .
-                *                                                            .
-                *                                                            .
-                *                                                        ]
-                *                                        
-                *                                  ]
+                * array must be set up as: 
+                * $array = [ 'custom_fields' => [
+                *                   ['display_name' => "Cart Id", "variable_name" => "cart_id", "value" => "2"],
+                *                   ['display_name' => "Sex", "variable_name" => "sex", "value" => "female"],
+                *                   .
+                *                   .
+                *                   .
+                *                  ]
+                *          ]
                 */
                 'metadata' => request()->metadata
-            ];
-
-            // Remove the fields which were not sent (value would be null)
-            array_filter($data);
+            ]);
         }
 
         $this->setHttpResponse('/transaction/initialize', 'POST', $data);
@@ -176,10 +189,10 @@ class Paystack
 
         return $this;
     }
-    
+
      /**
      * Get the authorization callback response
-     * In situations where Laravel serves as an backend for a detached UI, the api cannot redirect 
+     * In situations where Laravel serves as an backend for a detached UI, the api cannot redirect
      * and might need to take different actions based on the success or not of the transaction
      * @return array
      */
@@ -337,7 +350,8 @@ class Paystack
 
         $this->setRequestOptions();
 
-        $this->setHttpResponse("/plan", 'POST', $data);
+        return $this->setHttpResponse("/plan", 'POST', $data)->getResponse();
+
     }
 
     /**
@@ -449,7 +463,7 @@ class Paystack
         ];
 
         $this->setRequestOptions();
-        $this->setHttpResponse('/subscription', 'POST', $data);
+        return $this->setHttpResponse('/subscription', 'POST', $data)->getResponse();
     }
 
     /**
@@ -543,7 +557,7 @@ class Paystack
         ];
 
         $this->setRequestOptions();
-        $this->setHttpResponse('/page', 'POST', $data);
+        return $this->setHttpResponse('/page', 'POST', $data)->getResponse();
     }
 
     /**
@@ -582,5 +596,79 @@ class Paystack
 
         $this->setRequestOptions();
         return $this->setHttpResponse('/page/'.$page_id, 'PUT', $data)->getResponse();
+    }
+
+     /**
+     * Creates a subaccount to be used for split payments . Required    params are business_name , settlement_bank , account_number ,   percentage_charge
+     *
+     * @return array
+     */
+
+    public function createSubAccount(){
+        $data = [
+            "business_name" => request()->business_name,
+            "settlement_bank" => request()->settlement_bank,
+            "account_number" => request()->account_number,
+            "percentage_charge" => request()->percentage_charge,
+            "primary_contact_email" => request()->primary_contact_email,
+            "primary_contact_name" => request()->primary_contact_name,
+            "primary_contact_phone" => request()->primary_contact_phone,
+            "metadata" => request()->metadata,
+            'settlement_schedule' => request()->settlement_schedule
+        ];
+
+        $this->setRequestOptions();
+        return $this->setHttpResponse('/subaccount', 'POST', array_filter($data))->getResponse();
+
+    }
+
+     /**
+     * Fetches details of a subaccount
+     * @param subaccount code
+     * @return array
+     */
+    public function fetchSubAccount($subaccount_code){
+
+        $this->setRequestOptions();
+        return $this->setHttpResponse("/subaccount/{$subaccount_code}","GET",[])->getResponse();
+
+    }
+
+     /**
+     * Lists all the subaccounts associated with the account
+     * @param $per_page - Specifies how many records to retrieve per page , $page - SPecifies exactly what page to retrieve
+     * @return array
+     */
+    public function listSubAccounts($per_page,$page){
+
+        $this->setRequestOptions();
+        return $this->setHttpResponse("/subaccount/?perPage=".(int) $per_page."&page=".(int) $page,"GET")->getResponse();
+
+    }
+
+
+    /**
+     * Updates a subaccount to be used for split payments . Required params are business_name , settlement_bank , account_number , percentage_charge
+     * @param subaccount code
+     * @return array
+     */
+
+    public function updateSubAccount($subaccount_code){
+        $data = [
+            "business_name" => request()->business_name,
+            "settlement_bank" => request()->settlement_bank,
+            "account_number" => request()->account_number,
+            "percentage_charge" => request()->percentage_charge,
+            "description" => request()->description,
+            "primary_contact_email" => request()->primary_contact_email,
+            "primary_contact_name" => request()->primary_contact_name,
+            "primary_contact_phone" => request()->primary_contact_phone,
+            "metadata" => request()->metadata,
+            'settlement_schedule' => request()->settlement_schedule
+        ];
+
+        $this->setRequestOptions();
+        return $this->setHttpResponse("/subaccount/{$subaccount_code}", "PUT", array_filter($data))->getResponse();
+
     }
 }